Add files via upload
This program randomly generates numbers from 1 through 6 for two dice variables, dice1 and dice 2. The dices are rolled 3,600 times and the frequencies of the sum of the dices are recorded in an array. The sum of the two dices ranges from 2 - 12. After the dices are rolled, the frequencies of the sum of the two dices are displayed in a tabular form.