Reversal of elements in an array
This program takes an array and shifts the elements of the array to the left. The program allows the user to make the choice of the number of elements to shift to the left. In the program code, first two elements of the elements are reverse (or sent to the back).