Array Reversal
The code in this program rotates elements of an array to the left of the array. The programmer is allowed to specify the number of elements that should be rotated to the left of the array. In the code, the program rotates the first to elements of the array to the left, which causes them to appear at the back of the array.